# documentation/api.yaml
openapi: 3.0.0
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0
  licence:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: http://localhost:5000
    description: Dev Server on Port
  # - url: http://staging-api.example.com
  #   description: Optional server description, e.g. Internal staging server for testing
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
security:
  - bearerAuth: []
paths:
...
  /user/cuourseEnrole/{id}:
    put:
      summary: User Course enrole
      tags:
        - User Course
      description: We add the user to the course and we add the user ID to the course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to add to user cursors
          example: 62ae948249ef0d92e405b22b
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: integer
              example:
                userId: "62af2d6316c2ba27f61eb8df"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "User sucsess fuly enrolled cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/userCourseDelete/{id}:
    put:
      put:
      summary: User course delete
      tags:
        - User Course
      description: We add the user to the course and we add the user ID to the course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to add to user cursors
          example: 62ae948249ef0d92e405b22b
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: integer
              example:
                userId: "62af2d6316c2ba27f61eb8df"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                userEnroled: string
              example:
                userEnroled: "User sucsess fuly enrolled cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/profileUpdate:
    put:
      summary: Super Admin Update Profile
      tags:
        - User Course
      description: To change the Super Admin cabinet data, you must first enter the old password and the security word and then the new ones
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userId:
                  type: string
                name:
                  type: string
                email:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
                userAvatar:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                adminProfile: string
              example:
                adminProfile: Secsess fully updated User Profile
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/forgotPassword/searchUser:
    post:
      summary: User forget password update
      tags:
        - User Course
      description: When you forget the password, we confirm that the user is registered by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              example:
                email: Sunna@gmail.com
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                token: integer
                _id: integer
                name: string
                email: string
                password: string
                secretInfo: string
                createdAt: string
                updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/fotgetPassword/updatedPasword:
    post:
      summary: User forget password update
      tags:
        - User Course
      description: When you forget the password, we confirm that the user is registered by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                password:
                  type: string
              example:
                userId: 62af2d6316c2ba27f61eb8df
                password: "12345"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                secretInfo: string
                createdAt: string
                updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...